% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_abstracts_to_index_md.R
\name{add_abstracts_to_index_md}
\alias{add_abstracts_to_index_md}
\title{Add Abstracts to Hugo-Academic Publications}
\usage{
add_abstracts_to_index_md(endnote_df,
  hugo_pub_dir = "./content/publication")
}
\arguments{
\item{endnote_df}{endnote_df as retrieved by kwb.endnote::create_references_df()
or kwb.endnote::clean_references_df()}

\item{hugo_pub_dir}{hugo-academic folder containing publications
(default: "./content/publication")}
}
\value{
add abstracts to index.md
}
\description{
Add Abstracts to Hugo-Academic Publications
}
\examples{
\dontrun{
endnote_list <- kwb.endnote::create_endnote_list()
refs_df_clean <- kwb.endnote::clean_references_df(endnote_list)
add_abstracts_to_index_md(endnote_df = refs_df_clean)
}
}
