% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_authors_avatar.R
\name{add_author_avatar}
\alias{add_author_avatar}
\title{Add Author Avatar}
\usage{
add_author_avatar(
  author_metadata,
  overwrite = FALSE,
  hugo_root_dir = ".",
  width = 250,
  height = 300,
  x_off = 175,
  y_off = 0
)
}
\arguments{
\item{author_metadata}{one record as retrieved by add_authors_metadata()}

\item{overwrite}{default: FALSE}

\item{hugo_root_dir}{(default: ".")}

\item{width}{passed to magickx::geometry_area (default: 300)}

\item{height}{passed to magickx::geometry_area (default: 250)}

\item{x_off}{passed to magickx::geometry_area (default: 175)}

\item{y_off}{passed to magickx::geometry_area (default: 0)}
}
\value{
add avatar for author
}
\description{
Add Author Avatar
}
