% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_authors_index_md.R
\name{add_author_index_md}
\alias{add_author_index_md}
\title{Add Author Index.md}
\usage{
add_author_index_md(author_metadata, overwrite = FALSE,
  hugo_root_dir = ".")
}
\arguments{
\item{author_metadata}{one record as retrieved by add_authors_metadata()}

\item{overwrite}{default: FALSE}

\item{hugo_root_dir}{(default: ".")}
}
\value{
creates author index.md for one author
}
\description{
Add Author Index.md
}
