% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_authors_avatar.R
\name{add_authors_avatar}
\alias{add_authors_avatar}
\title{Add Authors Avatars}
\usage{
add_authors_avatar(authors_metadata = add_authors_metadata(),
  overwrite = FALSE, hugo_root_dir = ".")
}
\arguments{
\item{authors_metadata}{as retrieved by add_authors_metadata()}

\item{overwrite}{(default: FALSE)}

\item{hugo_root_dir}{(default: ".")}
}
\value{
creates author avatars for all authors
}
\description{
Add Authors Avatars
}
