% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_authors_index_md.R
\name{add_authors_index_md}
\alias{add_authors_index_md}
\title{Add Authors Index.md}
\usage{
add_authors_index_md(
  authors_metadata = add_authors_metadata(),
  overwrite = FALSE,
  hugo_root_dir = "."
)
}
\arguments{
\item{authors_metadata}{as retrieved by add_authors_metadata()}

\item{overwrite}{(default: FALSE)}

\item{hugo_root_dir}{(default: ".")}
}
\value{
creates author index.md for all authors
}
\description{
Add Authors Index.md
}
