% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_authors_avatar.R
\name{add_authors_metadata}
\alias{add_authors_metadata}
\title{Get Authors Metadata}
\usage{
add_authors_metadata(authors_config = get_authors_config())
}
\arguments{
\item{authors_config}{default: get_authors_config()}
}
\value{
add authors metadata (email, dirname)
}
\description{
Get Authors Metadata
}
