% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_columns_to_endnote_db.R
\name{add_columns_to_endnote_db}
\alias{add_columns_to_endnote_db}
\title{add_columns_to_endnote_db}
\usage{
add_columns_to_endnote_db(endnote_db_refs)
}
\arguments{
\item{endnote_db_refs}{table "refs" in Endnote DB (as retrieved by
[read_endnote_db])}
}
\value{
Endnote table with added columns using [add_publishdate_to_endnote_db]
and [add_kwb_style_to_reports]
}
\description{
add_columns_to_endnote_db
}
