% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_kwb-style_to_endnote-db.R
\name{add_kwb_style}
\alias{add_kwb_style}
\title{add_kwb_style}
\usage{
add_kwb_style(endnote_db_refs)
}
\arguments{
\item{endnote_db_refs}{table "refs" in Endnote DB (as retrieved by
[read_endnote_db])}
}
\value{
Endnotes "refs" table with added columns "reference_type_name"
(translating "reference_type" to "own" classes) and "publication" (used for
hugo-academic)
}
\description{
add_kwb_style
}
\examples{
\dontrun{
endnote_db <- read_endnote_db(path = "../../dms/2020-07-08/KWB-documents_20191205.Data/sdb/sdb.eni")
kwb_style_list <- add_kwb_style(endnote_db$refs)
}

}
