% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_projects_to_pub_index_md.R
\name{add_projects_to_pub_index_md}
\alias{add_projects_to_pub_index_md}
\title{Add Projects to Hugo-Academic Publications}
\usage{
add_projects_to_pub_index_md(
  endnote_df,
  overwrite = FALSE,
  col_project = "custom2",
  hugo_root_dir = "."
)
}
\arguments{
\item{endnote_df}{endnote_df as retrieved by kwb.endnote::create_references_df()
or kwb.endnote::clean_references_df()}

\item{overwrite}{should existing "projects" be overwritten (default: FALSE)}

\item{col_project}{name of column containing project id (default: "custom2")}

\item{hugo_root_dir}{root dir of hugo-academic website (default: ".")}
}
\value{
add project tags to index.md
}
\description{
Add Projects to Hugo-Academic Publications
}
\examples{
\dontrun{
endnote_list <- kwb.endnote::create_endnote_list()
refs_df_clean <- kwb.endnote::clean_references_df(endnote_list)
add_projects_to_pub_index_md(endnote_df = refs_df_clean)
}
}
