% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_abstracts_to_endnote_db.R,
%   R/add_publishdate_to_endnote_db.R
\name{add_publishdate_to_endnote_db}
\alias{add_publishdate_to_endnote_db}
\title{add_publishdate_to_endnote_db}
\usage{
add_publishdate_to_endnote_db(endnote_db_refs)

add_publishdate_to_endnote_db(endnote_db_refs)
}
\arguments{
\item{endnote_db_refs}{table "refs" in Endnote DB (as retrieved by
[read_endnote_db])}
}
\value{
Endnotes "refs" table with added columns "publish_datetime",
"publishDate" (required for hugo-academic) and "date-cleaned"

Endnotes "refs" table with added columns "publish_datetime",
"publishDate" (required for hugo-academic) and "date-cleaned"
}
\description{
add_publishdate_to_endnote_db

add_publishdate_to_endnote_db
}
\examples{
\dontrun{
endnote_db <- read_endnote_db(path = "../../dms/2020-07-08/KWB-documents_20191205.Data/sdb/sdb.eni")
en_db_with_publishdates <- add_publishdate_to_endnote_db(endnote_db$refs)
}
\dontrun{
endnote_db <- read_endnote_db(path = "../../dms/2020-07-08/KWB-documents_20191205.Data/sdb/sdb.eni")
en_db_with_publishdates <- add_publishdate_to_endnote_db(endnote_db$refs)
}
}
