% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_authors_index_md.R
\name{create_author_education}
\alias{create_author_education}
\title{Create Author Education}
\usage{
create_author_education(author_metadata)
}
\arguments{
\item{author_metadata}{one record as retrieved by add_authors_metadata()}
}
\value{
txt with education
}
\description{
Create Author Education
}
