% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_authors_index_md.R
\name{create_author_interests}
\alias{create_author_interests}
\title{Create Author Interests}
\usage{
create_author_interests(author_metadata)
}
\arguments{
\item{author_metadata}{one record as retrieved by add_authors_metadata()}
}
\value{
txt with interests
}
\description{
Create Author Interests
}
