% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_authors_index_md.R
\name{create_author_name}
\alias{create_author_name}
\title{Create Author Name}
\usage{
create_author_name(author_metadata)
}
\arguments{
\item{author_metadata}{one record as retrieved by add_authors_metadata()}
}
\value{
txt with author name
}
\description{
Create Author Name
}
