% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_authors_index_md.R
\name{create_author_social}
\alias{create_author_social}
\title{Create Author Social}
\usage{
create_author_social(author_metadata)
}
\arguments{
\item{author_metadata}{one record as retrieved by add_authors_metadata()}
}
\value{
txt with social icons
}
\description{
Create Author Social
}
