% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_authors_index_md.R
\name{create_author_user_groups}
\alias{create_author_user_groups}
\title{Create Author User Groups}
\usage{
create_author_user_groups(author_metadata)
}
\arguments{
\item{author_metadata}{one record as retrieved by add_authors_metadata()}
}
\value{
txt with user_groups
}
\description{
Create Author User Groups
}
