% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_code.R
\name{create_code}
\alias{create_code}
\title{Create Code}
\usage{
create_code(repo_topics, hugo_root_dir = ".")
}
\arguments{
\item{repo_topics}{data.frame as retrieved by pkgmeta::get_github_repos()}

\item{hugo_root_dir}{root dir of hugo-academic website (default: ".")}
}
\value{
create index.md for each project_name in subfolder "content/project"
}
\description{
Create Code
}
\examples{
\dontrun{
repo_topics <- pkgmeta::get_github_topics()
create_code(repo_topics = repo_topics)
}

}
