% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_projects.R
\name{create_projects}
\alias{create_projects}
\title{Create Projects}
\usage{
create_projects(project_names, hugo_root_dir = ".")
}
\arguments{
\item{project_names}{character vector with project acryonms to create
(as retrieved by kwb.endnote::clean_references_df()$custom2)}

\item{hugo_root_dir}{root dir of hugo-academic website (default: ".")}
}
\value{
create index.md for each project_name in subfolder "content/project"
}
\description{
Create Projects
}
\examples{
\dontrun{
endnote_list <- kwb.endnote::create_endnote_list()
endnote_df <- kwb.endnote::clean_references_df()
create_projects(project_names = endnote_df$custom2)
}
}
