% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_authors_index_md.R
\name{create_username}
\alias{create_username}
\title{Create Author Username}
\usage{
create_username(author_metadata)
}
\arguments{
\item{author_metadata}{one record as retrieved by add_authors_metadata()}
}
\value{
txt with author username
}
\description{
Create Author Username
}
