% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_authors_avatar.R
\name{get_authors_config}
\alias{get_authors_config}
\title{Get Authors Config}
\usage{
get_authors_config(path = config_authors_default())
}
\arguments{
\item{path}{path to authors config (default: config_authors_default())}
}
\value{
imports authors config as data.frame
}
\description{
Get Authors Config
}
