% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_publication_index_md_paths.R
\name{get_publication_index_md_paths}
\alias{get_publication_index_md_paths}
\title{get_publication_index_md_paths}
\usage{
get_publication_index_md_paths(hugo_root_dir = ".", lang = "")
}
\arguments{
\item{hugo_root_dir}{hugo root dir (default: .)}

\item{lang}{optional language code (default: ""), i.e publications will be
searched in <hugo_root_dir>/<language_code>/publication}
}
\value{
all recursive "index.md" files in folder
}
\description{
get_publication_index_md_paths
}
