% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_projects.R
\name{get_unique_project_names}
\alias{get_unique_project_names}
\title{Helper function: get unique project names}
\usage{
get_unique_project_names(project_names)
}
\arguments{
\item{project_names}{character vector with project acryonms to create
(as retrieved by kwb.endnote::clean_references_df()$custom2)}
}
\description{
Helper function: get unique project names
}
