% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_authors_index_md.R
\name{prepare_author_index_md}
\alias{prepare_author_index_md}
\title{Prepare Author "_index.md"}
\usage{
prepare_author_index_md(
  author_metadata,
  template_author_md = template_author_md_default()
)
}
\arguments{
\item{author_metadata}{one record as retrieved by add_authors_metadata()}

\item{template_author_md}{default: template_author_md_default()}
}
\value{
"_index.md" for author as text
}
\description{
Prepare Author "_index.md"
}
