% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_endnote_db.R
\name{read_endnote_db}
\alias{read_endnote_db}
\title{Read Endnote DB Tables}
\usage{
read_endnote_db(path)
}
\arguments{
\item{path}{full path to Endnote database ("sdb.eni")}
}
\value{
list with all tables contained in Endnote Databse
}
\description{
Read Endnote DB Tables
}
