% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace_author_umlauts_in_pub_index_md.R
\name{replace_authors_umlauts_in_pub_index_md}
\alias{replace_authors_umlauts_in_pub_index_md}
\title{Replace Author Umlauts in Multiple Publication Index.md}
\usage{
replace_authors_umlauts_in_pub_index_md(hugo_root_dir = ".",
  encoding = "UTF-8")
}
\arguments{
\item{hugo_root_dir}{root dir of hugo-academic website (default: ".")}

\item{encoding}{default: "UTF-8"}
}
\value{
rewrites "author" field in multiple "index.md" files without umlauts
but with "ae", "oe", "ue", "Ae", "Oe", "Ue", "ss" where needed
}
\description{
Replace Author Umlauts in Multiple Publication Index.md
}
