% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/replace_md_keys_with_endnote_db_vals_in_pub_index_md.R
\name{replace_dates_in_pub_index_md}
\alias{replace_dates_in_pub_index_md}
\title{replace_dates_in_pub_index_md}
\usage{
replace_dates_in_pub_index_md(
  md_paths,
  endnote_db_refs,
  md_key = "date",
  endnote_db_col = "date_cleaned",
  file_encoding = "UTF-8",
  escape_with_double_quote = FALSE,
  dbg = TRUE
)
}
\arguments{
\item{md_paths}{paths to "publication" markdown files}

\item{endnote_db_refs}{table "refs" in Endnote DB (as retrieved by
[add_columns_to_endnote_db])}

\item{md_key}{markdown key pattern to search for (default: "date")}

\item{endnote_db_col}{column with values to be used for replacement (default:
"date_cleaned")}

\item{file_encoding}{default: "UTF-8}

\item{escape_with_double_quote}{should be double quotes added "<endnote-value>" (default: FALSE)}

\item{dbg}{default: TRUE}
}
\value{
replaces "date" in markdown files based on values in column
defined with parameter "endnote_db_col"
}
\description{
replace_dates_in_pub_index_md
}
