% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace_umlauts.R
\name{replace_umlauts}
\alias{replace_umlauts}
\title{Replace Umlauts}
\usage{
replace_umlauts(string)
}
\arguments{
\item{string}{string for replacement}
}
\value{
returns string without umlauts but with "ae", "oe", "ue",
"Ae", "Oe", "Ue" where needed
}
\description{
Replace Umlauts
}
