% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_lines.R
\name{write_lines}
\alias{write_lines}
\title{Helper function: write Lines To File with File Encoding}
\usage{
write_lines(text, file, fileEncoding = "", ...)
}
\arguments{
\item{text}{text}

\item{file}{file}

\item{fileEncoding}{fileEncoding}

\item{...}{additional arguments passed to [writeLines]}
}
\value{
writes file with "fileEncoding"
}
\description{
Helper function: write Lines To File with File Encoding
}
