% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_code.R
\name{create_code}
\alias{create_code}
\title{Create Code}
\usage{
create_code(repo_names, hugo_root_dir = ".")
}
\arguments{
\item{repo_names}{character vector with code repo acryonms to create
(as retrieved by pkgmeta::get_github_repos())}

\item{hugo_root_dir}{root dir of hugo-academic website (default: ".")}
}
\value{
create index.md for each project_name in subfolder "content/project"
}
\description{
Create Code
}
\examples{
\dontrun{
repo_topics <- pkgmeta::get_github_topics()
repo_names <- stringr::str_replace_all(stringr::str_remove_all(unique(repo_topics$full_name),
".*/"), "\\\\.", "-")
create_code(repo_names = repo_names)
}
}
