% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace_author_umlauts_in_pub_index_md.R
\name{get_publication_index_md_paths}
\alias{get_publication_index_md_paths}
\title{Get Publication Index.md Paths}
\usage{
get_publication_index_md_paths(hugo_root_dir = ".")
}
\arguments{
\item{hugo_root_dir}{root dir of hugo-academic website (default: ".")}
}
\value{
vector with absolute paths with all publication "index.md" files
}
\description{
Get Publication Index.md Paths
}
